+++ src/main/java/fr/openent/evaluations/service/DevoirService.java
@@ -65,7 +65,7 @@ public interface DevoirService extends CrudService {
      * @param idPeriode identifiant de la période
      * @param handler handler portant le résultat de la requête
      */
-    void listDevoirs(String idEtablissement, String idClasse, String idMatiere, Integer idPeriode, Handler<Either<String, JsonArray>> handler);
+    void listDevoirs(String idEtablissement, String idClasse, String idMatiere, Long idPeriode, Handler<Either<String, JsonArray>> handler);
 
     /**
      * Liste des devoirs publiés pour un établissement et une période donnée.
@@ -76,7 +76,7 @@ public interface DevoirService extends CrudService {
      * @param idUser identifant de l'utilisateur
      * @param handler handler portant le résultat de la requête
      */
-    void listDevoirs(String idEtablissement, Integer idPeriode, String idUser,Handler<Either<String, JsonArray>> handler);
+    void listDevoirs(String idEtablissement, Long idPeriode, String idUser,Handler<Either<String, JsonArray>> handler);
 
     /**
      * Récupère le nombre de notes en fonction du devoir pour un utilisateur donné
